requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP requests for testing
requests==2.31.0

# Development tools
python-multipart==0.0.6
python-dotenv==1.0.0

# OCR and image processing
pytesseract==0.3.10
Pillow==10.1.0
PyMuPDF==1.23.8
opencv-python==4.8.1.78

# Document processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
unidecode==1.3.7

# LangChain and AI
groq==0.4.1
langchain>=0.1.0
langchain-groq>=0.0.1
langchain-core>=0.1.7
langchain-community>=0.0.10